Corregido el problema para bajas Tr en CS y VdW. El inconveniente principal estaba en la convergencia del metodo de Newton. Para solucionarlo, se implemento una funcion que reduce sistematicamente el valor de Vr0 inicial, hasta que se invierta el signo de la diferencia de \'areas. Es un poco m\'as lento, pero robusto. El m\'etodo de Newton queda como opcional

+++ VdWEos.py
@@ -6,31 +6,31 @@ class VdWEos:
     "Ecuaci\'on de estado de Van der Waals"
 
     def __init__(self,a,b):
-        self.__a = a
-        self.__b = b
+        self.__a = np.float64(a)
+        self.__b = np.float64(b)
         pass
 
     def Pr(self,T,v):
         "Presi\'on reducida"
         
-        return 8.*T/(3.*v - 1.) - 3./v**2
+        return np.float64( 8.*T/(3.*v - 1.) - 3./v**2 )
 
 
     def Tc(self):
         "Temperatura cr\'itica"
 
-        return 8. * self.__a / (27. * self.__b)
+        return np.float64( 8. * self.__a / (27. * self.__b) )
 
 
     def Pc(self):
         "Presi\'on cr\'itica"
 
-        return self.__a / (27. * self.__b**2)
+        return np.float64( self.__a / (27. * self.__b**2) )
 
     def rhoc(self):
         "Densidad cr\'itica"
 
-        return 1. / (3. * self.__b)
+        return np.float64( 1. / (3. * self.__b) )
     
     
     def __str__(self):
